refactor: rename data/models→entities, lift AppDatabase.ts (mirrors legado)
Match legado Android structure:
  data/models/ → data/entities/   (io.legado.app.data.entities)
  data/database/AppDatabase.ts → data/AppDatabase.ts

Update all imports across dao/, app/, src/core/network/.
