+++ src/data/entities/ReadConfig.ts
